pantry search now uses only selected ingredients. Also combined search and pantry components
