+++ src/Routes.js
@@ -11,13 +11,12 @@ import Recipe from "./components/Recipe";
 import { Route, Routes } from "react-router-dom";
 import LandingPage from "./components/LandingPage";
 import Navbar from "./components/Navbar/Navbar";
-import SearchResults from "./components/RecipeResults/SearchResults";
+import RecipeSearch from "./components/RecipeResults/RecipeSearch";
 import { me, getPantries, loadIngredients, loadShoppingList } from "./store";
 import { useDispatch, useSelector } from "react-redux";
 import AccountPage from "./components/account/AccountPage";
 import CreateRecipe from "./components/CreateRecipe";
 import ShoppingList from "./components/account/ShoppingList";
-// import PantryList from "./components/pantry/PantryList";
 import SavedRecipes from "./components/account/SavedRecipes";
 
 
@@ -52,7 +51,7 @@ const ClientRoutes = () => {
         <Route path="recipe/:id" element={<Recipe />} />
         <Route path="account" element={<AccountPage />} />
         <Route path='/shoppingList' element={<ShoppingList/>} />
-        <Route path="searchResults" element={<SearchResults />} />
+        <Route path="searchResults" element={<RecipeSearch />} />
         <Route path="create-recipe" element={<CreateRecipe />} />
         <Route path='savedRecipes' element={<SavedRecipes />}/>
       </Routes>
